# archionApi/src/api/core/models/archives.model.ts
export class Archives {
    
    id?: number;
    UNID?: string;
    CommityNameAndNum?: string;
    CommityName?: string;
    CommityNumber?: number;
    CommityDecsionDate?: string;
    ForumDescription1?: string;
    ForumDescription?: string;
    GovDecision?: number;
    GovDecisionDate?: string;
    DecisionTitle?: string;
     OpeningText?: string;
     DecisionText?: string;
     AppealText?: string;
     DecisionType?: string;
     DecisionNumAndPrimeName?: string;
     SecriteryNum?: number;
     Embeded?: string;
	
}